refactor: Move constants out to a separate module
